+++ BlazorMonaco/Bridge/Data.cs
@@ -1,0 +1,34 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace BlazorMonaco.Bridge
+{
+    public class Position
+    {
+        public int LineNumber { get; set; }
+        public int Column { get; set; }
+    }
+
+    public class Range
+    {
+        public int StartLineNumber { get; set; }
+        public int StartColumn { get; set; }
+        public int EndLineNumber { get; set; }
+        public int EndColumn { get; set; }
+    }
+
+    public class Selection : Range
+    {
+        public int SelectionStartLineNumber { get; set; }
+        public int SelectionStartColumn { get; set; }
+        public int PositionLineNumber { get; set; }
+        public int PositionColumn { get; set; }
+    }
+
+    public class Dimension
+    {
+        public int Width { get; set; }
+        public int Height { get; set; }
+    }
+}