+++ BlazorMonaco/Bridge/FindOptions.cs
@@ -1,9 +1,0 @@
-﻿namespace BlazorMonaco.Bridge
-{
-    public class FindOptions
-    {
-        public bool? AddExtraSpaceOnTop { get; set; }
-        public bool? AutoFindInSelection { get; set; }
-        public bool? SeedSearchStringFromSelection { get; set; }
-    }
-}