+++ BlazorMonaco/Bridge/HoverOptions.cs
@@ -1,9 +1,0 @@
-﻿namespace BlazorMonaco.Bridge
-{
-    public class HoverOptions
-    {
-        public int? Delay { get; set; } = 300;
-        public bool? Enabled { get; set; } = true;
-        public bool? Sticky { get; set; } = true;
-    }
-}