+++ BlazorMonaco/Bridge/LightbulbOptions.cs
@@ -1,7 +1,0 @@
-﻿namespace BlazorMonaco.Bridge
-{
-    public class LightbulbOptions
-    {
-        public bool? Enabled { get; set; }
-    }
-}