+++ BlazorMonaco/Bridge/ParameterHintOptions.cs
@@ -1,8 +1,0 @@
-﻿namespace BlazorMonaco.Bridge
-{
-    public class ParameterHintOptions
-    {
-        public bool? Cycle { get; set; }
-        public bool? Enabled { get; set; } = true;
-    }
-}