+++ docs/_content/BlazorMonaco/jsInterop.js
@@ -11,18 +11,17 @@ window.blazorMonaco.editor = {
     },
 
     create: function (id, options) {
-        if (window.blazorMonaco.editors.find(e => e.id === model.id)) {
+        if (window.blazorMonaco.editors.find(e => e.id === model.id))
             return;
-        }
         
-        require(['vs/editor/editor.main'], function () {
-            if (options == null) {
-                options = {};
-            }
+        if (options == null)
+            options = {};
 
-            var editor = monaco.editor.create(document.getElementById(id), options);
-            window.blazorMonaco.editors.push({ id: id, editor: editor });
-        });
+        if (typeof monaco === 'undefined')
+            console.log("WARNING : Please check that you have the script tag for editor.main.js in your index.html file");
+
+        var editor = monaco.editor.create(document.getElementById(id), options);
+        window.blazorMonaco.editors.push({ id: id, editor: editor });
     },
 
     setTheme: function (theme) {
@@ -36,12 +35,10 @@ window.blazorMonaco.editor = {
 
     getEditorById: function (id) {
         let editorHolder = window.blazorMonaco.editors.find(e => e.id === id);
-        if (!editorHolder) {
+        if (!editorHolder)
             throw "Couldn't find the editor with id: " + id + " editors.length: " + window.blazorMonaco.editors.length;
-        }
-        else if (!editorHolder.editor) {
+        else if (!editorHolder.editor)
             throw "editor is null for editorHolder: " + editorHolder;
-        }
         return editorHolder.editor;
     },
 
@@ -71,6 +68,36 @@ window.blazorMonaco.editor = {
     getValue: function (id) {
         let editor = this.getEditorById(id);
         return editor.getValue();
+    },
+
+    setEventListener: function (id, eventName, handler) {
+        let editor = this.getEditorById(id);
+
+        let listener = function (e) {
+            handler.invokeMethodAsync("EventCallback", eventName, JSON.stringify(e));
+        };
+
+        switch (eventName) {
+            case "OnContextMenu": editor.onContextMenu(listener); break;
+            case "OnDidBlurEditorText": editor.onDidBlurEditorText(listener); break;
+            case "OnDidBlurEditorWidget": editor.onDidBlurEditorWidget(listener); break;
+            case "OnDidChangeConfiguration": editor.onDidChangeConfiguration(listener); break;
+            case "OnDidChangeCursorPosition": editor.onDidChangeCursorPosition(listener); break;
+            case "OnDidChangeCursorSelection": editor.onDidChangeCursorSelection(listener); break;
+            case "OnDidContentSizeChange": editor.onDidContentSizeChange(listener); break;
+            case "OnDidDispose": editor.onDidDispose(listener); break;
+            case "OnDidFocusEditorText": editor.onDidFocusEditorText(listener); break;
+            case "OnDidFocusEditorWidget": editor.onDidFocusEditorWidget(listener); break;
+            case "OnDidLayoutChange": editor.onDidLayoutChange(listener); break;
+            case "OnDidPaste": editor.onDidPaste(listener); break;
+            case "OnDidScrollChange": editor.onDidScrollChange(listener); break;
+            case "OnKeyDown": editor.onKeyDown(listener); break;
+            case "OnKeyUp": editor.onKeyUp(listener); break;
+            case "OnMouseDown": editor.onMouseDown(listener); break;
+            case "OnMouseLeave": editor.onMouseLeave(listener); break;
+            case "OnMouseMove": editor.onMouseMove(listener); break;
+            case "OnMouseUp": editor.onMouseUp(listener); break;
+        }
     },
 
     setValue: function (id, value) {