+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Change Log
 
+## 1.1.0
+* Added support for editor events.
+* Changed editor GetOptions parameter name to ConstructionOptions.
+* Bug fixes
+
 ## 1.0.3
 * Updated for Blazor v3.2.0-preview5
 
 ## 1.0.2
-* Bug fixes
+* Bug fixes.
 
 ## 1.0.1
-* Initial release
+* Initial release.